update prop definitions to match the most recent version of the library
